numcodecs_pco

Enum PcodecError

pub enum PcodecError {
    UnsupportedDtype(AnyArrayDType),
    HeaderEncodeFailed {
        source: PcoHeaderError,
    },
    PcoEncodeFailed {
        source: PcoCodingError,
    },
    EncodedDataNotBytes {
        dtype: AnyArrayDType,
    },
    EncodedDataNotOneDimensional {
        shape: Vec<usize>,
    },
    HeaderDecodeFailed {
        source: PcoHeaderError,
    },
    PcoDecodeFailed {
        source: PcoCodingError,
    },
    DecodeInvalidShapeHeader {
        source: ShapeError,
    },
    MismatchedDecodeIntoArray {
        source: AnyArrayAssignError,
    },
}

Errors that may occur when applying the Pcodec.

Variants

UnsupportedDtype(AnyArrayDType)

Pcodec does not support the dtype

HeaderEncodeFailed

Pcodec failed to encode the header

Fields

source: PcoHeaderError

Opaque source error

PcoEncodeFailed

Pcodec failed to encode the data

Fields

source: PcoCodingError

Opaque source error

EncodedDataNotBytes

Pcodec can only decode one-dimensional byte arrays but received an array of a different dtype

Fields

dtype: AnyArrayDType

The unexpected dtype of the encoded array

EncodedDataNotOneDimensional

Pcodec can only decode one-dimensional byte arrays but received an array of a different shape

Fields

shape: Vec<usize>

The unexpected shape of the encoded array

HeaderDecodeFailed

Pcodec failed to decode the header

Fields

source: PcoHeaderError

Opaque source error

PcoDecodeFailed

Pcodec failed to decode the data

Fields

source: PcoCodingError

Opaque source error

DecodeInvalidShapeHeader

Pcodec decoded an invalid array shape header which does not fit the decoded data

Fields

source: ShapeError

Source error

MismatchedDecodeIntoArray

Pcodec cannot decode into the provided array

Fields

source: AnyArrayAssignError

The source of the error