numcodecs_pco

Function compress

pub fn compress<T: PcoElement, S: Data<Elem = T>, D: Dimension>(
    data: ArrayBase<S, D>,
    level: PcoCompressionLevel,
    mode: PcoModeSpec,
    delta: PcoDeltaSpec,
    paging: PcoPagingSpec,
) -> Result<Vec<u8>, PcodecError>

Compresses the input data array using pco with the given compression level, mode, delta encoding, and paging mode.

Errors

Errors with

PcodecError::HeaderEncodeFailed if encoding the header failed

PcodecError::PcoEncodeFailed if encoding failed with an opaque error