numcodecs_pco

Function decompress_into

pub fn decompress_into<T: PcoElement, D: Dimension>(
    encoded: &[u8],
    decoded: ArrayViewMut<'_, T, D>,
) -> Result<(), PcodecError>

Decompresses the encoded data into the decoded array.

Errors

Errors with

PcodecError::HeaderDecodeFailed if decoding the header failed

PcodecError::MismatchedDecodeIntoArray if the decoded array has the wrong dtype or shape

PcodecError::PcoDecodeFailed if decoding failed with an opaque error

PcodecError::DecodeInvalidShapeHeader if the array shape header does not fit the decoded data